Handle Strings (slightly) better (#106)
Still no escapes etc.

And we should decide whether we want to do single vs. double quoted.
When I originally wrote the proposal I defaulted to double, but I
realize that some SQL engines use double quotes for some specific
things, so possibly we need to retain those for that? Or force people to
escape them.