Refine rust code
Making the rust a bit nicer & reducing cloning a bit.

I'm not going to invest too much here (and some of it is for my own
learning in areas I'd skipped over in my rust education...). I _think_
most of the code is fairly legible (let me know if not though, very open
to feedback). And while the parser is the "oldest" part of the the code,
there are other parts of the code which need more work — I'm trying to
iterate between making stuff work and keeping the code legible.