Overhaul parsing of nested objects (#132)
* Overhaul parsing of nested objects

This is in preparation for a better approach to running functions. I
realize there are lots of changes — more than I'd like at this stage,
but it's mostly an internal change to the parser.

I've commented out the test of running functions since that will change anyway.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>