Remove link break in badges (#306)
* Add link to reference docs

I've set up a domain prql.builders, and linked to the language docs at https://lang.prql.builders.

`.builders` is not perfect. The other one that was reasonable was `prql.tools`. We could do `prql`rs`, but its main users would be PRQL users rather than rust developers (and it's a minor pain to buy the domain since it's actually Serbian).

* Remove link break in badges