Standardize test-databases crate
So we're not special-casing things in GitHub Actions which would cause local tests to fail. One way to check that is whether `task test-all` works.

(Hope it's OK to do this @aljazerzen -- I thought easier to do it here given I'd done it before in other crates.)

BTW one option would be to make this an integration test in `prql-compiler`. I don't have much experience with the best way of organizing; but at least the promise of cargo integration tests is a separate compilation unit.

We could also try and integrate with our existing examples (and maybe those should go into an examples unit in prql-compiler...)