Add Changelog
I think the idea is that for anything material, we endeavor to add a changelog entry. If people tire of it, then we can fall back to me writing one at release time -- more important to get contributions than keep a pretty changelog, but people do like reading changelogs (me included!).

We can use cargo-cliff to automate these, but that requires a bit more work, and getting it started will help.