Add changelog for next version
I've bumped the minor version here, to 0.3.0. This might be a bit
conservative, because I see the stdlib issue as more of a bug than a
change.

But maybe people like @mklopets are relying on that?

(I guess 0.1.0 & 0.2.0 were both big releases, and this is very small.
But they don't _have_ to be coupled)