Refactor FormattedError
Not the most essential code change — and open to feedback on whether
it's useful on net! — but trying to make the interface simpler:
- Moved related functions into a Trait
- Renamed the Trait containing the error message from `FormattedError` to `ErrorMessage`, since it's not an
  `Error` type (maybe there are better alternatives though?)

Still not confident about the overall interface. Should `compile` return
a `Result` whose `Error` type has the full message on? At that point we
have `source`, albeit not the `source_id`.