refactor: FormattedError (#1182)
* tests: An initial error message test

Hopefully we can expand on these; maybe they could be an interesting way for folks to get started.

I left a comment / question in the code about the `FormattedError` struct

* Refactor FormattedError

Not the most essential code change — and open to feedback on whether
it's useful on net! — but trying to make the interface simpler:
- Moved related functions into a Trait
- Renamed the Trait containing the error message from `FormattedError` to `ErrorMessage`, since it's not an
  `Error` type (maybe there are better alternatives though?)

Still not confident about the overall interface. Should `compile` return
a `Result` whose `Error` type has the full message on? At that point we
have `source`, albeit not the `source_id`.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>