build: Add `cli` as a default feature
This reverts a change from @aljazerzen, so I'll wait for his review. It's required for `cargo install` to work without a `--features` flag.

Even if we don't want this, I would vote to merge this, release 0.4.1, and then do the work to the docs & brew recipe so that works again