revert: "feat: Code snippets from book in Playground (#2070)" (#2144)
This is causing the build to break and so we can't deploy 0.6.1 to the web. It's completely my fault; we should have had a test for this; ironically I added an issue for this yesterday https://github.com/PRQL/prql/issues/2122.

I just spent 20 mins trying to fix the issue, but couldn't manage, and need to log off for the night, so reverting this as per our standard approach. I'm hoping it won't be a difficult fix and we can re-revert this and add it back easily.

The error is:

```
src/sidebar/Sidebar.js
  Line 30:10:  'v' is assigned a value but never used  no-unused-vars
```