refactor: Implement Ident as arbitrary list
Part of #1535, this changes our Idents in PL from `$namespace.$name` to an arbitrary hierachy.

It's full of `.clone()` & `.into()` -- I've been doing this by replacing one definition and then adding lots of `.into()` untils it passes (not the most conceptual work!). Doing it incrementally at least means I can't end up in a quagmire of not knowing why the new version doesn't work; it's easy to revert to the last known good state. We can do another pass to improve the rust / reduce allocations.

The plan here would be to:
- replace any other usages in the compilation prior to the ident being resolved. Once it's resolved, it doesn't necessarily need a full hierarchy. Possibly we can remove the old `Ident` / rename the new one to `Ident`.
- adjust how the resolution works so we can have arbitrary hierarchies of schema (`a.b.c`). Still some work to think about how this should work (some initial comments in #1535)
- make backticks fully opaque, so #1535 works -- then using parquet files will be easy, we won't need to quote schemas, the semantics will be simple & consistent