fix: Don't panic on `group (take)`
I can't really work out how to do this well — I'm guessing that the error should be handled long before it gets to the point where it's trying to do the `take` on the group?

I asked a couple of questions on Discord re what a `Ty` / `SetExpr` was.

I previously mentioned that RQ was really promising in its ability to reduce the amount of context that was required to develop on the compiler. We briefly touched on whether there were additional IRs that would be useful. I don't think we necessarily need a full IR, but I think it would be useful to have a more structured representation of the intermediate AST that we could use to debug errors like this — is there some stage that this gets to which is correct beyond PL, where we can see what's obviously wrong, and we could show the relevant Struct?