docs(changelog): fix typo and tweak some items related to breaking chage of 0.6.0 (#2366)
Related to #2104

A set of some minor fixes that I noticed when I looked at the Changelog
that is now included in the Book.

- Fix some typos.
- Added some info on some of breaking changes
  - Pass through jinja template (`{{` and `}}`).
  - `switch` -> `case`

I tried to include the admonish note about the change of name from
`switch` to `case` in the changelog, but it didn't work (bug?).
So I add it to the `case` page.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
(cherry picked from commit ca23ad9ba727500e48cb5fe494af4f44e0fcc41c)