docs: Another approach for combining Expressions & Parentheses (#2460)
I, too, merged the "Expressions & Operators" page with the "Precedence
and Parentheses" page. It's remarkable about how similar the result is
to #2455.

I was never sold on the notion that "precedence" was the most important
concept, so I re-wrote the rules to describe when parentheses were
necessary.

There are many embedded questions that need to be addressed before this
could be merged. I would appreciate your thoughts.

@max-sixty @aljazerzen @eitsupi

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Maximilian Roos <5635139+max-sixty@users.noreply.github.com>