docs: Move Terminology
I'm not sure where this should go, but its current place I don't think is that coherent any longer.

I also removed the index in the book for Internals — I think better to have no index than an out-of-date one. Would be great if there were a tool to build this automatically!