fix: compile_help test from prqlc
The test introduced in 0d6481633a0be823ed7880af85e16de067031df1 never
worked because a pre-commit hook had automatically stripped the trailing
whitespace. This wasn't spotted in the CI because cargo-insta 1.30.0 has
a bug that it returns exit code 0 even if there's a test failure.[1]

This commit restores the necessary trailing whitespace and disables the
trailing-whitespace pre-commit hook (we already run `cargo fmt` anyway,
which is intelligent enough to preserve trailing whitespace in strings).

[1]: https://github.com/mitsuhiko/insta/issues/392