refactor: extract prql_ast crate
There are a bunch of reasons why it makes sense to have the actual
abstract syntax tree (AST) in a dedicated crate, e.g:

* the code formatter just needs the AST, it doesn't need type
  information (and the previous mixing up of syntax elements
  with inferred info was very confusing ... case in point the
  "codegen" formatter even formatted inferred args for
  Pipeline syntax nodes).

* Currently there's just the prql-compiler which does PRQL -> SQL,
  but the other direction SQL -> PRQL is also interesting,
  and you should be able to implement mappings to PRQL syntax elements
  without depending on the compiler or the parser.

* Some crates may even want to interpret PRQL queries
  directly without SQL being involved at all
  (e.g. using PRQL as a simple filtering language or
  perhaps even linting PRQL queries.)

So this commit decouples a prql_ast crate by making the Expr struct
generic over a new trait with a bunch of associated types.
The new types look roughly as follows:

    pub struct Expr<T: Extension> {
        pub kind: ExprKind<T>,
        pub span: Option<Span>,
        pub alias: Option<String>,

        #[serde(flatten)]
        pub extra: T::ExprExtra,
    }

    pub enum ExprKind<T: Extension> {
        Ident(Ident),
        Literal(Literal),
        Pipeline(Pipeline<T>),
        Tuple(Vec<Expr<T>>),
        Array(Vec<Expr<T>>),
        // etc.

        Other(T::ExprKindVariant),
    }

To prevent a bunch of unreadable snapshot test diffs that would make it
hard to verify that this change hasn't broken anything, this commit
adds a custom Serialize impl for ExprKind that will be removed in the
following commit.

(Note that we add this Serialize impl in lib.rs because adding it in
expr.rs would prevent the git file rename detection at the default 50%
similarity threshold.)