refactor!: remove impl From<ExprKind> for Expr
From impls should be used for cases where the conversion is obvious.
Here we are also initializing a bunch of other fields with default
values, making a constructor function more apt.

Switching to a constructor function also enables us to change the
function to accept `impl Into<ExprKind>` in the future, which wouldn't
be as cleanly achievable with Into since .into().into() fails with
"type annotations needed" and we should avoid blanket implementations
if possible to keep the API simple.