refactor: move Span to ast module
We want to split the AST and the parser into their own crates.
The parser crate would depend on the AST crate. Spans are part
of the AST so they will have to be part of the AST crate
(unless we'd make the AST generic over the span type but there's
really no need for that).

Note that we intentionally don't move the Add<usize> and chumsky::Span
impls to the ast module since they are specific to the parser (once we
split the crates the parser will have to introduce a newtype around Span
from the AST to still be able to implement these traits for these types).