refactor: move Range helpers out of impl
We want to move Range along with a bunch of other types to a new
prql_ast crate. Since you cannot directly write an impl block for a
foreign type, this commit (and the following commits) move a bunch of
methods for types we want to extract out of impl blocks.