refactor: move codegen::test to tests::test_codegen
The code formatting tests rely on the parser, which for now remains in
this crate, so we cannot move the tests along with the codegen module to
prql_ast.

The plan is to split both the parser as well as the code formatter into
separate crates in the future, at which point the tests will again be
able to live happily close to the respective code.

Doing this move in a separate commit to help the git rename detection.