refactor: newtype Span in parser
We want to move the Span type to the prql_ast crate. If we do that,
the parser can however no longer implement the chumsky::Span trait for
prql_ast::Span (since you cannot implement a foreign trait for a foreign
type). So this commit introduces a newtype around Span, so that we can
avoid moving this parser-specific chumsky::Span impl to prql_ast.