feat: partial application
When a function call argument is a function, wrap the call into a new function
which expects the missing argument of the argument function.

This is equialent to changing this:

```
let a = (sort x | take 1)
```
... into this:
```
let a = rel -> (rel | sort x | take 1)
```

We already supported this, but only for pipelines.
This PR generalizes this behavior to plain function calls.

The downside is that when your function *does expect* a function as an
argument, this behavior will get in the way. In this case, the user can to
annotate the paramater to be a function, with the convenient recent addition
of `<func>` type annotation.

PS: this behavior may not be called partial application, I'm not sure and too
lazy to look it up.