ci: Revert some of #3551
I'm not sure what I was thinking with some of the changes in #3551 -- clearly if we only discriminate by `run_number`, we're not going to cancel an earlier run when a later run gets pushed to the same branch / ref.

We saw this this morning when dependabot added labels -- the earlier runs didn't get canceled, and we had a long queue.