tests: Add tests for Lexer output
My overall goal is to output each stage of the compiler pipeline for a set of canonical queries.

I started with the lexer output. I had to make more changes than I expected — to get `Display` on the tokens, we needed to add a couple of custom structs and implement `Display` on those.

Maybe there's a less invasive way?