ci: Only run on macos-14
These are gradually transitioning to `macos-latest`, but because they can't share a cache, we need to tag them explicitly. When it's fully transitioned, we can specify `latest` again (would be great if Dependabot could upgrade runner images too so we could upgrade explicitly!)

Based on https://github.blog/changelog/2024-04-01-macos-14-sonoma-is-generally-available-and-the-latest-macos-runner-image/

This means we no longer test on intel cpus on MacOS. But seems that's the way things are going.