refactor: Use a `parse_with_parser` func to test any parser
We haven't been able to just add a test for any parser so far; so we've only tested full exprs or full stmts. This adds a helper functions, and uses it to create the existing `parse_expr` and `parse_single` functions.

It also moves the `expr` tests down a level.