internal: Add a `Start` token at the beginning of LR output
This is required as we add some aesthetics to the parser and need to handle newlines. If we want to enforce anything starting on a new line (for example a doc comment), then we need to handle the start of a file, which is semantically a new line but doesn't have a literal new line token. So we add this token and then will allow it to act as a new line.