refactor: Use `Option<Box<T>>`
I was trying to understand more about `.as_deref` and used prql code as a learning ground. Apparently `Option<Box<T>>` is supposed to be more idiomatic than `Box<Option<T>>`. This makes a few changes in that direction (mostly as a practcial exercise for me, but hopefully a marginal improvement in our code too).